Config and cwa CHADA tables
Updated __init__.py to include a function to be called
from the command line as an entry point, which generates
a config file for this package specifying the location
of the plantuml.jar. Other config parameters may be added
later. __init__.py also finds the path to the 'themes'
directory, containing the bespoke CHADA themes supplied.
This directory is two levels above the __file__ of __init__.py.

The modules tables.py and running.py contain functions for
generating plantuml code files and the resulting diagrams from
the contents of a single json file. The json file should be
formatted according to the template CHADA_TABLES_TEMPLATE.json
and this is similar to the table format in the current working
agreement of the European Materials Characterisation Committee.
Having formatted the json file, the resulting tables including
hyperlinks if desiered, can be generated by calling a single
Python function: make_chada_tables().