Minor theme updates
Updated the docstring for the theme MOCHADA-plasma_A4w.puml

Added a new theme: MOCAHDA-CWA_A4ww.puml. This theme is intended
for the production of individual CHADA/MODA tables without aborts
contents element, which can be pasted into a document. The width
of the elements is optimised ofr A4 paper and the background
colour of the elements is white.