Version bump to 0.1.2, added tables/copy_theme_to_local_folder()
Previously when making CHADA tbales using the functions in
mochada_kit/tables, you could either set the path for the
themes to be relative to the mochada_kit/gallery folder
or put the full path to mochada_kit/themes in the file.
I have now added a third option, which is to copy the
requested theme to a themes folder located in the same
directory and set the relative path to that folder
in the puml code. This avoids having full paths in puml
code, which it may not be safe to share online.